requirements.txt
pandas
numpy
openai
tensorflow
scikit-learn
llama_index
langchain
pdfminer.six
simpletransformers
flask
PyPDF2
torch
matplotlib

# -e .